+++ sdk/resources/azure-mgmt-resource-deployments/azure/mgmt/resource/deployments/_configuration.py
@@ -34,7 +34,7 @@ class DeploymentsMgmtClientConfiguration:  # pylint: disable=too-many-instance-a
     :param cloud_setting: The cloud setting for which to get the ARM endpoint. Default value is
      None.
     :type cloud_setting: ~azure.core.AzureClouds
-    :keyword api_version: The API version to use for this operation. Known values are "2025-04-01"
+    :keyword api_version: The API version to use for this operation. Known values are "2026-06-01"
      and None. Default value is None. If not set, the operation's default API version will be used.
      Note that overriding this default value may result in unsupported behavior.
     :paramtype api_version: str
@@ -48,7 +48,7 @@ class DeploymentsMgmtClientConfiguration:  # pylint: disable=too-many-instance-a
         cloud_setting: Optional["AzureClouds"] = None,
         **kwargs: Any
     ) -> None:
-        api_version: str = kwargs.pop("api_version", "2025-04-01")
+        api_version: str = kwargs.pop("api_version", "2026-06-01")
 
         if credential is None:
             raise ValueError("Parameter 'credential' must not be None.")