+++ sdk/resources/azure-mgmt-resource-deployments/generated_samples/calculate_template_hash.py
@@ -49,6 +49,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2025-04-01/CalculateTemplateHash.json
+# x-ms-original-file: 2026-06-01/CalculateTemplateHash.json
 if __name__ == "__main__":
     main()