+++ sdk/resources/azure-mgmt-resource-deployments/generated_samples/post_deployment_what_if_on_tenant.py
@@ -44,6 +44,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2025-04-01/PostDeploymentWhatIfOnTenant.json
+# x-ms-original-file: 2026-06-01/PostDeploymentWhatIfOnTenant.json
 if __name__ == "__main__":
     main()